Move the `wcdl/section` directory into the `components` directory and rename related CSS.
