+++ js/src/components/free-listings/choose-audience-section/choose-audience-section.js
@@ -8,7 +8,7 @@ import { __ } from '@wordpress/i18n';
  */
 import { useAdaptiveFormContext } from '~/components/adaptive-form';
 import AppRadioContentControl from '~/components/app-radio-content-control';
-import Section from '~/wcdl/section';
+import Section from '~/components/section';
 import Subsection from '~/wcdl/subsection';
 import RadioHelperText from '~/components/radio-helper-text';
 import SupportedCountrySelect from '~/components/supported-country-select';
