+++ js/src/components/free-listings/configure-product-listings/shipping-time-setup/index.js
@@ -7,7 +7,7 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import { useAdaptiveFormContext } from '~/components/adaptive-form';
-import Section from '~/wcdl/section';
+import Section from '~/components/section';
 import AppSpinner from '~/components/app-spinner';
 import ShippingCountriesForm from './shipping-countries-form';
