+++ js/src/components/google-ads-account-card/claim-account/index.js
@@ -8,7 +8,7 @@ import { Fragment } from '@wordpress/element';
  * Internal dependencies
  */
 import { useAppDispatch } from '~/data';
-import Section from '~/wcdl/section';
+import Section from '~/components/section';
 import useWindowFocusCallbackIntervalEffect from '~/hooks/useWindowFocusCallbackIntervalEffect';
 import DisconnectAccount from '../disconnect-account';
 import './index.scss';
