+++ js/src/components/google-ads-account-card/connect-ads/index.js
@@ -14,7 +14,7 @@ import AppDocumentationLink from '~/components/app-documentation-link';
 import ConnectButton from './connect-button';
 import ContentButtonLayout from '~/components/content-button-layout';
 import LoadingLabel from '~/components/loading-label';
-import Section from '~/wcdl/section';
+import Section from '~/components/section';
 import Subsection from '~/wcdl/subsection';
 import useApiFetchCallback from '~/hooks/useApiFetchCallback';
 import useDispatchCoreNotices from '~/hooks/useDispatchCoreNotices';
