+++ js/src/components/google-ads-account-card/connected-google-ads-account-card.js
@@ -9,7 +9,7 @@ import { ExternalLink } from '@wordpress/components';
 import toAccountText from '~/utils/toAccountText';
 import AccountCard, { APPEARANCE } from '~/components/account-card';
 import ConnectedIconLabel from '~/components/connected-icon-label';
-import Section from '~/wcdl/section';
+import Section from '~/components/section';
 import DisconnectAccount from './disconnect-account';
 
 /**
