+++ js/src/components/paid-ads/ads-campaign/paid-ads-features-section.scss
@@ -3,7 +3,7 @@
 		color: $gray-700;
 	}
 
-	.wcdl-section-card-title {
+	.gla-section-card-title {
 		margin-bottom: $grid-unit-15;
 		line-height: $gla-line-height-medium-large;
 		font-weight: normal;
