+++ js/src/components/paid-ads/budget-section/index.scss
@@ -1,5 +1,5 @@
 .gla-budget-section {
-	.wcdl-section__body {
+	.gla-section__body {
 		display: flex;
 		flex-direction: column;
 		gap: $grid-unit-20;
