+++ js/src/components/paid-ads/campaign-preview/campaign-preview-card.js
@@ -10,7 +10,7 @@ import GridiconChevronRight from 'gridicons/dist/chevron-right';
 /**
  * Internal dependencies
  */
-import Section from '~/wcdl/section';
+import Section from '~/components/section';
 import AppButton from '~/components/app-button';
 import CampaignPreview from './campaign-preview';
 import './campaign-preview-card.scss';
