+++ js/src/components/paid-ads/campaign-preview/campaign-preview-card.scss
@@ -1,5 +1,5 @@
 .gla-campaign-preview-card {
-	.wcdl-section-card-title {
+	.gla-section-card-title {
 		margin-bottom: $grid-unit-10;
 	}
