+++ js/src/components/section/card/body.scss
@@ -1,3 +1,3 @@
-.wcdl-section-card-body {
+.gla-section-card-body {
 	padding: var(--large-gap);
 }
