+++ js/src/components/section/card/footer.scss
@@ -1,4 +1,4 @@
-.wcdl-section-card-footer {
+.gla-section-card-footer {
 	padding: calc(var(--large-gap) / 2) var(--large-gap);
 
 	&[hidden] {
