+++ js/src/components/section/card/title.js
@@ -14,7 +14,7 @@ const Title = ( props ) => {
 
 	return (
 		<Subsection.Title
-			className={ classnames( 'wcdl-section-card-title', className ) }
+			className={ classnames( 'gla-section-card-title', className ) }
 			{ ...rest }
 		/>
 	);
