+++ js/src/components/section/card/title.scss
@@ -1,3 +1,3 @@
-.wcdl-section-card-title {
+.gla-section-card-title {
 	margin-bottom: calc(var(--main-gap) / 3 * 2);
 }
