+++ js/src/components/section/index.scss
@@ -1,4 +1,4 @@
-.wcdl-section {
+.gla-section {
 	display: flex;
 	flex-direction: column;
 	margin-bottom: var(--large-gap);
@@ -33,7 +33,7 @@
 		}
 	}
 
-	.wcdl-section__body {
+	.gla-section__body {
 		flex: 1;
 	}
 }
