+++ js/src/components/spinner-card/index.js
@@ -2,7 +2,7 @@
  * Internal dependencies
  */
 import AppSpinner from '~/components/app-spinner';
-import Section from '~/wcdl/section';
+import Section from '~/components/section';
 
 const SpinnerCard = () => {
 	return (
