+++ js/src/components/stepper/step-content-footer/index.js
@@ -1,7 +1,7 @@
 /**
  * Internal dependencies
  */
-import Section from '~/wcdl/section';
+import Section from '~/components/section';
 import VerticalGapLayout from '~/components/vertical-gap-layout';
 
 const StepContentFooter = ( { children } ) => {
