+++ js/src/pages/onboarding/setup-stepper/setup-accounts/index.js
@@ -17,7 +17,7 @@ import StepContent from '~/components/stepper/step-content';
 import StepContentHeader from '~/components/stepper/step-content-header';
 import StepContentFooter from '~/components/stepper/step-content-footer';
 import StepContentActions from '~/components/stepper/step-content-actions';
-import Section from '~/wcdl/section';
+import Section from '~/components/section';
 import AppDocumentationLink from '~/components/app-documentation-link';
 import VerticalGapLayout from '~/components/vertical-gap-layout';
 import WPComAccountCard from '~/components/wpcom-account-card';
