+++ js/src/pages/settings/index.scss
@@ -1,5 +1,5 @@
 .gla-settings {
-	.wcdl-section {
+	.gla-section {
 		max-width: 780px;
 		margin-left: auto;
 		margin-right: auto;
