+++ js/src/pages/settings/reconnect-google-account/disconnect-google-account-card.js
@@ -13,7 +13,7 @@ import { useAppDispatch } from '~/data';
 import useAdminUrl from '~/hooks/useAdminUrl';
 import AppButton from '~/components/app-button';
 import AccountCard, { APPEARANCE } from '~/components/account-card';
-import Section from '~/wcdl/section';
+import Section from '~/components/section';
 import DisconnectModal, { ALL_ACCOUNTS } from '../disconnect-modal';
 
 export default function DisconnectGoogleAccountCard( { email } ) {
