+++ js/src/pages/settings/reconnect-google-account/reconnect-google-account.js
@@ -12,7 +12,7 @@ import { glaData } from '~/constants';
 import { getDashboardUrl } from '~/utils/urls';
 import toScopeState from '~/utils/toScopeState';
 import useAppSelectDispatch from '~/hooks/useAppSelectDispatch';
-import Section from '~/wcdl/section';
+import Section from '~/components/section';
 import AppSpinner from '~/components/app-spinner';
 import GoogleAccountCard from '~/components/google-account-card';
 import DisconnectGoogleAccountCard from './disconnect-google-account-card';
